+++ src/wsdd.py
@@ -1115,11 +1115,11 @@ class ApiServer:
         # Thus, we ignore type errors here.
         if isinstance(listen_address, int) or listen_address.isnumeric():
             self.server = await aio_loop.create_task(asyncio.start_server(  # type: ignore
-                self.on_connect, host='localhost', port=int(listen_address), loop=aio_loop, reuse_address=True,
+                self.on_connect, host='localhost', port=int(listen_address), reuse_address=True,
                 reuse_port=True))
         else:
             self.server = await aio_loop.create_task(asyncio.start_unix_server(  # type: ignore
-                self.on_connect, path=listen_address, loop=aio_loop))
+                self.on_connect, path=listen_address))
 
     async def on_connect(self, read_stream: asyncio.StreamReader, write_stream: asyncio.StreamWriter) -> None:
         while True: